Split CombiningAdapter unit tests into 3 tests
